refactor(dis_html): move CVTERM/config-coupled HTML helpers out of dis_responder
With config constants now in dis_config and CVTERM/PROJECT in dis_state (prior two
commits), three presentation helpers no longer need the Flask app and move to
dis_html.py: add_jrc_fields, get_license, add_subjects. They import CVTERM from
dis_state and DO_NOT_DISPLAY/NCBI_MESH/PMCID from dis_config, and are re-exported
via `from dis_html import ...` so call sites are unchanged.

Not moved (not pure presentation): get_display_badges and add_orcid_controls do
network calls (JRC.call_oa / requests.get); compute_preprint_data returns data,
not HTML; generate_works_table / add_orcid_works depend on get_work_* accessors
that stay in dis_responder. Those belong in a future data/fetch module, not here.

No behavior change. Verified: both modules parse and import; pyflakes reports zero
undefined names in each; the moved helpers read CVTERM populated in dis_state
(dis_responder.CVTERM is dis_state.CVTERM); the DOI detail page and the
/doiui/custom CVTERM canary both return 200. Patch bump to 119.24.4.
